Factored out DSL code in to own module
This makes the "main" class a bit more readable.
